feat: update start operator field design
