+++ packages/toolkit/src/lib/use-instill-form/components/index.ts
@@ -1,2 +1,4 @@
 export * from "./regular";
 export * from "./component-output";
+export * from "./smart-hint";
+export * from "./start-operator-free-form-fields";
