+++ packages/toolkit/src/lib/use-instill-form/components/start-operator-free-form-fields/index.ts
@@ -1,0 +1,21 @@
+import { AudioField } from "./AudioField";
+import { AudiosField } from "./AudiosField";
+import { BooleanField } from "./BooleanField";
+import { ImageField } from "./ImageField";
+import { ImagesField } from "./ImagesField";
+import { NumberField } from "./NumberField";
+import { NumbersField } from "./NumbersField";
+import { TextField } from "./TextField";
+import { TextsField } from "./TextsField";
+
+export const StartOperatorFreeFormFields = {
+  AudioField,
+  AudiosField,
+  BooleanField,
+  ImageField,
+  ImagesField,
+  NumberField,
+  NumbersField,
+  TextField,
+  TextsField,
+};
