+++ packages/toolkit/src/lib/use-instill-form/index.ts
@@ -1,3 +1,4 @@
 export * from "./type";
 export * from "./useComponentOutputFields";
 export * from "./useInstillForm";
+export * from "./useStartOperatorFreeForm";
