+++ packages/toolkit/src/lib/use-instill-form/transform/index.ts
@@ -3,3 +3,5 @@ export * from "./transformInstillFormTreeToDefaultValue";
 export * from "./transformInstillFormTreeToInitialSelectedCondition";
 export * from "./transformInstillJSONSchemaToFormTree";
 export * from "./transformInstillJSONSchemaToZod";
+export * from "./transformStartOperatorMetadataToSuperRefineRules";
+export * from "./transformStartOperatorMetadataToZod";
