+++ packages/toolkit/src/view/pipeline-builder/components/end-operator-node/UserDefinedFieldItem.tsx
@@ -13,13 +13,13 @@ const selector = (store: InstillStore) => ({
 });
 
 export const UserDefinedFieldItem = ({
-  outputKey,
-  outputValue,
+  fieldKey,
+  fieldValue,
   onDeleteField,
   onEditField,
 }: {
-  outputKey: string;
-  outputValue: string;
+  fieldKey: string;
+  fieldValue: string;
   onDeleteField: (key: string) => void;
   onEditField: (key: string) => void;
 }) => {
@@ -32,7 +32,7 @@ export const UserDefinedFieldItem = ({
     transform,
     transition,
     isDragging,
-  } = useSortable({ id: outputKey });
+  } = useSortable({ id: fieldKey });
 
   const reactFlowInstance = useReactFlow();
 
@@ -53,7 +53,7 @@ export const UserDefinedFieldItem = ({
 
   return (
     <div
-      key={outputKey}
+      key={fieldKey}
       className={cn(
         "nodrag group flex cursor-default flex-row gap-x-2 bg-semantic-bg-base-bg",
         isDragging ? "z-10" : ""
@@ -76,25 +76,25 @@ export const UserDefinedFieldItem = ({
       <div className="flex w-full flex-col">
         <div className="mb-2 flex flex-row items-center justify-between">
           <div className="my-auto font-sans text-base font-semibold text-semantic-fg-primary">
-            {outputKey}
+            {fieldKey}
           </div>
           {currentVersion === "latest" && isOwner ? (
             <div className="my-auto flex flex-row gap-x-2">
               <button
-                onClick={() => onEditField(outputKey)}
+                onClick={() => onEditField(fieldKey)}
                 className="flex flex-row gap-x-1 rounded-full bg-semantic-accent-bg px-2 py-0.5 font-sans text-xs font-medium text-semantic-accent-default"
               >
                 edit field
                 <Icons.Edit03 className="my-auto h-3 w-3 stroke-semantic-accent-on-bg" />
               </button>
-              <button onClick={() => onDeleteField(outputKey)}>
+              <button onClick={() => onDeleteField(fieldKey)}>
                 <Icons.Trash01 className="h-3 w-3 stroke-semantic-error-on-bg" />
               </button>
             </div>
           ) : null}
         </div>
         <p className="nodrag rounded-sm border border-semantic-bg-line bg-semantic-bg-primary px-1.5 py-[9px] product-body-text-4-regular">
-          {outputValue}
+          {fieldValue}
         </p>
       </div>
       {/* Placeholder */}
