+++ packages/toolkit/src/view/pipeline-builder/components/start-node/index.ts
@@ -1,1 +1,1 @@
-export * from "./StartNode";
+export * from "./StartOperatorNode";
