+++ packages/toolkit/src/view/pipeline-builder/index.ts
@@ -1,6 +1,5 @@
 export * from "./components";
 export * from "./lib";
-export * from "./use-node-input-fields";
 export * from "./Flow";
 export * from "./FlowControl";
 export * from "./RightPanel";
