--- packages/toolkit/src/view/pipeline-builder/use-node-input-fields/index.ts
@@ -1,1 +1,0 @@
-export * from "./useStartOperatorTestModeInputForm";
